+++ server/schemas/user.py
@@ -1,17 +1,17 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, EmailStr
 from datetime import datetime
 
 class UserBase(BaseModel):
-    email: str
+    email: EmailStr
     password: str
     class Config:
         orm_mode = True
 
-
 class CreateUser(UserBase):
     pass
 
-class UserResponse(UserBase):
+class UserResponse(BaseModel):
+    email: EmailStr
     id : int
     created_at: datetime
     class Config:
